requirements.txt
pyrogram==2.0.106
pydantic==1.10.18
requests==2.32.3
fastapi==0.110.3
uvicorn==0.30.3
sentry-sdk==2.17.0
fpdf2==2.7.9
pillow==10.4.0

